move FromBase64Int to core
