+++ core.go
@@ -1,6 +1,7 @@
 package cose
 
 import (
+	"encoding/base64"
 	"bytes"
 	"crypto"
 	"crypto/ecdsa"
@@ -230,4 +231,16 @@ func I2OSP(b *big.Int, n int) []byte {
 		return buf.Bytes()
 	}
 	return os[:n]
+}
+
+// FromBase64Int decodes a base64-encoded string into a big.Int or panics
+//
+// from https://github.com/square/go-jose/blob/789a4c4bd4c118f7564954f441b29c153ccd6a96/utils_test.go#L45
+// Apache License 2.0
+func FromBase64Int(data string) *big.Int {
+	val, err := base64.RawURLEncoding.DecodeString(data)
+	if err != nil {
+		panic("Invalid test data")
+	}
+	return new(big.Int).SetBytes(val)
 }
